Install the basic package independent of the basiclite package.
The basic and the basiclite packages are incompatible from now because
they install same files except NLS data.

The sqlplus formula depends on the basic pakage by default from
now. If you need the basiclite package, install it with the
--with-basiclite option as follows.

   brew install instantclient-sqlplus --with-basiclite

I chose the basic package as the default dependent package because the
basic package contains the basiclite package and I think that the
former is less problematic than the latter. For example sqlplus using
the latter cannot connect to an Oracle server when the database
character set is not supported by basiclite.